First successful build
