+++ src/routes/api/ratelimit/+server.ts
@@ -1,0 +1,7 @@
+export async function GET(request: Request) {
+    return new Response("rate limit exceeded", {status: 429});
+}
+
+export async function POST(request: Request) {
+    return new Response("rate limit exceeded", {status: 429});
+}
